+++ src/batou/tests/test_endtoend.py
@@ -48,15 +48,15 @@ ERROR: Failed override attribute conversion
  Attribute: Component1.do_what_is_needed
 Conversion: convert_literal('false')
      Error: malformed node or string: <_ast.Name object at 0x...>
-ERROR: Overrides for undefined attributes
-      Host: localhost
- Component: Component2
-Attributes: this_does_not_exist
 ERROR: Failed override attribute conversion
       Host: localhost
  Attribute: DNSProblem.attribute_with_problem
 Conversion: Address('localhost')
      Error: Need port for service address.
+ERROR: Overrides for undefined attributes
+      Host: localhost
+ Component: Component2
+Attributes: this_does_not_exist
 ERROR: Unused provided resources
     Resource "backend" provided by component3 with value ['192.168.0.1']
     Resource "frontend" provided by component3 with value ['test00.gocept.net']